\name{calcABC}
\alias{calcABC}
\title{Function to calculate ABC and OFL from projection model output.}
\usage{
calcABC(res = NULL, fn = NULL, pstar = 0.49, buffer = 0.1,
  title = NULL)
}
\arguments{
  \item{res}{- r list from TCSAM2013 projection model run}

  \item{fn}{- filename of TCSAM2013 projection model run r
  file}

  \item{pstar}{-}

  \item{buffer}{-}

  \item{title}{- title for plot}
}
\description{
This function calculates ABC and OFL from TCSAM2013
projection model output.
}
\details{

}

