% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParams.R
\name{checkParams}
\alias{checkParams}
\title{Function to plot parameter values.}
\usage{
checkParams(obj.prs = NULL, obj.std = NULL, dp = 0.01)
}
\arguments{
\item{obj.prs}{- TCSAM2013 prs object or 'active' or 'all'}

\item{obj.std}{- dataframe obtained from reading, or filename of, the .std file}

\item{dp}{- percent difference between parameter value and upper/lower limits used to color plot}
}
\value{
- dataframe
}
\description{
This function plots parameters values, together with their limits
(if any) and the posterior distributions implied by their estimated standard
errors.
}
\details{
If obj.std is NULL, the user will NOT be prompted to select an
std file and std info will NOT be included in the output.
}

