% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.FisheryQuantities.R
\name{compareModelResults.FisheryQuantities}
\alias{compareModelResults.FisheryQuantities}
\title{Compare estimated/predicted survey quatities among several model runs}
\usage{
compareModelResults.FisheryQuantities(obj, numRecent = 15,
  plot1stObs = TRUE, showPlot = FALSE, pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{numRecent}{- number of recent years to plot}

\item{plot1stObs}{- flag to plot observations from the first case, only}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- name for output pdf file}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare estimated/predicted survey quantities among 
several model runs.
}
\details{
Uses \code{getMDFR.FisheryQuantities}.
}

