% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.FisheryTCFMaleSelFcns.R
\name{compareModelResults.FisheryTCFMaleSelFcns}
\alias{compareModelResults.FisheryTCFMaleSelFcns}
\title{Compare estimated/predicted TCF male selectivity functions among several model runs}
\usage{
compareModelResults.FisheryTCFMaleSelFcns(obj, showPlot = FALSE, pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- name for output pdf file}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare estimated/predicted TCF male selectivity functions among 
several model runs.
}
\details{
Uses \code{getMDFR.FisheryQuantities}.
}
