% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.GrowthMatrices.R
\name{compareModelResults.GrowthMatrices}
\alias{compareModelResults.GrowthMatrices}
\title{Compare growth matrices among several model runs}
\usage{
compareModelResults.GrowthMatrices(obj, showPlot = FALSE, pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{showPlot}{- flag to print plot to current device}

\item{pdf}{- name for output pdf file}
}
\value{
list of ggplot objects
}
\description{
Function to compare growth matrices among several model runs.
}
\details{
None.
}

