% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.MeanGrowth.R
\name{compareModelResults.MeanGrowth}
\alias{compareModelResults.MeanGrowth}
\title{Compare mean post-molt growth among several model runs}
\usage{
compareModelResults.MeanGrowth(obj, showPlot = FALSE, pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{showPlot}{- flag to print plot to current device}

\item{pdf}{- name for output pdf file}
}
\value{
ggplot object
}
\description{
Function to compare mean post-molt growth among several model runs.
}
\details{
None.
}
