% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.NatMort.R
\name{compareModelResults.NatMort}
\alias{compareModelResults.NatMort}
\title{Function to plot natural mortality rates by year using ggplot2}
\usage{
compareModelResults.NatMort(obj, showPlot = TRUE, pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{showPlot}{- flag (T/F) to show plot}
}
\value{
ggplot2 object
}
\description{
This function plots natural mortality estimates by year
  and sex for immatures, new shell matures and old shell matures.
}
\details{
None.
}

