% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.ObjFunComponents.R
\name{compareModelResults.ObjFunComponents}
\alias{compareModelResults.ObjFunComponents}
\title{Compare changes in objective function components from a base model for several models}
\usage{
compareModelResults.ObjFunComponents(obj, base = NULL, y = c("diff",
  "objFun"), showPlot = FALSE, pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{base}{- name of case to use as base model, or NULL to use first}

\item{y}{- 'diff' or 'objFun'}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- filename for pdf output}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare changes in objective function components 
from a base model for several models.
}
\details{
If 'base' is not given, objective function component comparisons are relative to the first case. NEGATIVE
differences indicate a better fit (smaller value for the component in the alternative model). Selecting 
\code{y="diff"} plots relative objective function values--i.e., relative to the base model. Selecting \code{y="objFun"}
plots absolute values.
}
