% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.SelFcns.R
\name{compareModelResults.SelFcns}
\alias{compareModelResults.SelFcns}
\title{Compare selectivity functions among several model runs.}
\usage{
compareModelResults.SelFcns(objs = NULL, cases = NULL, dodge_width = NULL,
  dodge_height = NULL, pdf = "ModelComparisons.SelFcns")
}
\arguments{
\item{objs}{- list of objects derived from Jack's R files for the models to be compared}

\item{cases}{- vector of labels for model cases (if 'objs' is not given)}

\item{pdf}{- name for output pdf file}

\item{width}{- width for pdf output}

\item{height}{- height for pdf output}
}
\value{
vector of list objects corresponding to the objects returned by each model R file.
}
\description{
Function to compare selectivity functions among 
several model runs.
}
\details{
If 'objs' is not given, then the user is prompted to select Jack's R file output from each 
model to be compared. If 'cases' is given, the user is prompted to select the file
corresponding to each case. If 'cases' is not given, then the user may select an 
arbitrary number of files (one at a time), ending selection by pressing 'cancel' on the selection box'.\cr\cr
If 'objs' is not given, the working directory is set two levels above the 1st model case file selected.\cr\cr
Uses \code{PBSmodelling::readList}, \code{reshape2::melt} and \code{wtsUtilities::selectFile}.
}

