% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.SurveyPRs.R
\name{compareModelResults.SurveyPRs}
\alias{compareModelResults.SurveyPRs}
\title{Compare Pearson's residuals from survey size comps among several model runs}
\usage{
compareModelResults.SurveyPRs(obj = NULL, numRecent = 15,
  plot1stObs = TRUE, ncol = 5, showPlot = FALSE, pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{numRecent}{- number of recent years to plot}

\item{plot1stObs}{- flag to plot observations from the first case, only}

\item{ncol}{- number of columns/page for size comp plots}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- name for output pdf file}
}
\value{
non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare Pearson's residuals from  survey size comps among 
several model runs.
}
\details{
Uses \code{getMDFR.SurveyQuantities}.
}

