% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.SurveySelFcns.R
\name{compareModelResults.SurveySelFcns}
\alias{compareModelResults.SurveySelFcns}
\title{Compare estimated/predicted survey selectivity functions among several model runs}
\usage{
compareModelResults.SurveySelFcns(obj = NULL, showPlot = FALSE,
  pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- name for output pdf file}
}
\value{
non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare estimated/predicted survey selectivity functions among 
several model runs.
}
\details{
Uses \code{getMDFR.SurveyQuantities}.
}

