% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.SuveyTimeSeries.R
\name{compareModelResults.SurveyTimeSeries}
\alias{compareModelResults.SurveyTimeSeries}
\title{Compare estimated/predicted survey time series among several model runs}
\usage{
compareModelResults.SurveyTimeSeries(reps = NULL, numRecent = 15,
  styr = NULL, endyr = NULL, plot1stObs = TRUE, plotLegal = FALSE,
  obsyr = NULL, pltyr = NULL,
  pdf = "ModelComparisons.SurveyTimeSeries.pdf")
}
\arguments{
\item{reps}{- list of objects derived from Jack's R files for the models to be compared}

\item{numRecent}{- number of recent years to plot}

\item{styr}{- start year for model run}

\item{endyr}{- final year of model run (assessment year)}

\item{plot1stObs}{- flag to plot observations from the first case, only}

\item{plotLegal}{- flag to plot legal abundance}

\item{obsyr}{- start year for survey observations}

\item{pltyr}{- start year for plots with only model-predicted values}

\item{pdf}{- name for output pdf file}

\item{reclag}{- recruitment lag}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare estimated/predicted survey time series among 
several model runs.
}
\details{
Uses \code{getMDFR.SurveyTimeSeries}.
}

