\name{exportModelComparisons.TimeSeries}
\alias{exportModelComparisons.TimeSeries}
\title{Export time series from a set of model runs to csv for comparison.}
\usage{
exportModelComparisons.TimeSeries(obsyrs = NULL, obs = NULL, obscv = NULL,
  prdyrs = NULL, vartype = NULL, objs = NULL, scaleBy = 1)
}
\description{
Function to export time series from a set of model runs to
csv for comparison.
}
\details{

}

