% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getActiveParams.R
\name{getActiveParams}
\alias{getActiveParams}
\title{Create a TCSAM2013 prs object from an active params file.}
\usage{
getActiveParams(in.prs = NULL)
}
\arguments{
\item{in.prs}{= filename of active params file}
}
\value{
list object corresponding to the active params file, or NULL if file does not exist
}
\description{
Function to create a TCSAM2013 prs object from an active params file.
}
\details{

}

