% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fisheries.CatchBiomass.R
\name{getMDFR.Fisheries.CatchBiomass}
\alias{getMDFR.Fisheries.CatchBiomass}
\title{Get fishery catch biomass time series from model results from TCSAM2013 model runs as a dataframe}
\usage{
getMDFR.Fisheries.CatchBiomass(obj, category = c("captured", "discarded",
  "retained", "discard mortality", "total mortality", "index"), cast = "y+x",
  fleets = c("TCF", "SCF", "RKF", "GTF"), verbose = FALSE)
}
\arguments{
\item{obj}{- single tcsam2013.rep object, tcsam2013.resLst object, or named list of the latter}

\item{category}{- 'captured','discarded','retained', 'discard mortality', 'total mortality', or 'index'}

\item{cast}{- casting formula (e.g., "x+m") for excluding x,m,s,z factor levels from a sum across the unspecified factors}

\item{fleets}{- name(s) of fisheries}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get fishery catch biomass time series from model results from TCSAM2013 model runs as a dataframe.
}
\details{
Extracts the estimated fishery biomass time series.
}

