% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.FisheryQuantities.R
\name{getMDFR.FisheryQuantities}
\alias{getMDFR.FisheryQuantities}
\title{Get estimated/predicted fishery-related quantities from several model runs}
\usage{
getMDFR.FisheryQuantities(obj, type = c("bio.retm", "bio.dscm", "bio.totm",
  "prNatZ.ret", "prNatZ.tot", "PRs.ret", "PRs.tot", "mnPrNatZ.ret",
  "mnPrNatZ.tot", "selfcns", "zscores", "effSS.ret", "effSS.tot", "max rates",
  "mean rates"), verbose = FALSE)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{type}{- fishery-related quantity to retrieve}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get estimated/predicted fishery-related quantities from 
several model runs.
}
\details{
Potential values for 'type' are:
\itemize{
 \item {'bio.retm' - retained catch mortality (1000's t)}
 \item {'bio.dscm' - discard catch mortality (1000's t) for males in directed fishery}
 \item {'bio.totm' - total catch mortality (1000's t)}
 \item {'prNatZ.ret' - annual retained catch proportions-at-size}
 \item {'prNatZ.tot' - annual total catch proportions-at-size}
 \item {'PRs.ret'      - pearsons residuals for retained catch proportions-at-size}
 \item {'PRs.tot'      - pearsons residuals for total catch proportions-at-size}
 \item {'mnPrNatZ.ret' - mean proportions-at-size for retained catch data}
 \item {'mnPrNatZ.tot' - mean proportions-at-size for total catch data}
 \item {'selfcns' - fishery selectivity and retention functions, by time period and sex}
 \item {'zscores' - annual z-scores for fits to catch biomass}
 \item {'effSS.tot' - effective (and input) sample sizes for total catch size comps}
 \item {'effSS.ret' - effective (and input) sample sizes for retained catch size comps}
 \item {'max rates' - max fishing mortality, retained mortality, and capture rates}
 \item {'mean rates' - mean fishing mortality, retained mortality, and capture rates}
}
Requires sqldf package.
}

