% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.MatureBiomass.R
\name{getMDFR.MatureBiomass}
\alias{getMDFR.MatureBiomass}
\title{Get predicted mature biomass (time series) from several model runs}
\usage{
getMDFR.MatureBiomass(obj, verbose = FALSE)
}
\arguments{
\item{obj}{- object with results for the models to be compared that can be converted to a list of tcsam2013.resLst objects}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get predicted mature biomass (time series) from 
several model runs.
}
\details{
Uses \code{reshape2::melt} and \code{reshape2::dcast}. If obj is a tcsam2013.lst
object, then results will be extracted from the tcsam2013.std object for a given case, if it exists; 
otherwise they will be extracted from the tcsam2013.rep object.
}

