% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ObjFunComponents.R
\name{getMDFR.ObjFunComponents}
\alias{getMDFR.ObjFunComponents}
\title{Get changes to objective function components relative to a base model for several models}
\usage{
getMDFR.ObjFunComponents(obj, base = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{base}{- name of case to use as base model, or NULL to use first}
}
\value{
dataframe
}
\description{
Function to get changes to objective function components 
relative to a base model for several models.
}
\details{
Objective function component comparisons are relative to the base (or first) case. NEGATIVE
differences indicate a better fit (smaller value for the component in the alternative model).
}
