% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ParamsPlusStdDevs.R
\name{getMDFR.ParamsPlusStdDevs}
\alias{getMDFR.ParamsPlusStdDevs}
\title{Function to extract parameter estimates and standard deviations from different TCSAM2013 models}
\usage{
getMDFR.ParamsPlusStdDevs(obj, dp = 0.01, verbose = FALSE)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{dp}{- percent difference between parameter value and upper/lower limits used to flag outliers}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
- dataframe with the parameter estimates and uncertainty info.
}
\description{
This function extracts parameters values and standard deviations, together with their limits
(if any) from several TCSAM2013 models.
}
\details{
Returned parameter info dataframe has columns case, type, param, min, max, init, value, stdvv and scl.
}

