% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.Abundance.R
\name{getMDFR.Pop.Abundance}
\alias{getMDFR.Pop.Abundance}
\title{Get population abundance time series from model results from TCSAM2013 model runs as a dataframe}
\usage{
getMDFR.Pop.Abundance(obj, cast = "x", verbose = FALSE)
}
\arguments{
\item{obj}{- single tcsam2013.rep object, tcsam2013.resLst object, or named list of the latter}

\item{cast}{- casting formula (e.g., "x+m") for excluding x,m,s,z factor levels from a sum of N_yxmsz across the unspecified factors}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get population abundance time series from model results from TCSAM2013 model runs as a dataframe.
}
\details{
Extracts the estimated population abundance time series.
}
