% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.Quantities.R
\name{getMDFR.Pop.Quantities}
\alias{getMDFR.Pop.Quantities}
\title{Get predicted population quantities (time series) from several model runs}
\usage{
getMDFR.Pop.Quantities(obj, type = c("MB_yx", "R_y", "iN_xmsz", "fN_xmsz",
  "B_yxmsz", "B_yxmz", "B_yxms", "B_yxm", "B_yx", "N_yxmsz", "N_yxmz", "N_yxms",
  "N_yxm", "N_yx"), verbose = FALSE)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{type}{- population quantity to retrieve}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe
}
\description{
Function to get predicted population quantities (time series) from 
several model runs.
}
\details{
Potential values for 'type' are:
\itemize{
 \item {'MB_yx' - mature biomass at spawning time (1000's t)}
 \item {'R_y' - recruitment (millions)}
 \item {'iN_xmsz' - initial (July 1) abundance by x,m,s,z (millions)}
 \item {'fN_xmsz' - final (July 1) abundance by x,m,s,z (millions)}
 \item {'B_yxmsz' - annual (July 1) abundance by x,m,s,z (millions)}
 \item {'B_yxmz'  - annual (July 1) abundance by x,m,z (millions)}
 \item {'B_yxz'   - annual (July 1) abundance by x,z (millions)}
 \item {'B_yxms'  - annual (July 1) abundance by x,m,s (millions)}
 \item {'B_yxm'   - annual (July 1) abundance by x,m (millions)}
 \item {'B_yx'    - annual (July 1) abundance by x (millions)}
 \item {'N_yxmsz' - annual (July 1) abundance by x,m,s,z (millions)}
 \item {'N_yxmz'  - annual (July 1) abundance by x,m,z (millions)}
 \item {'N_yxz'   - annual (July 1) abundance by x,z (millions)}
 \item {'N_yxms'  - annual (July 1) abundance by x,m,s (millions)}
 \item {'N_yxm'   - annual (July 1) abundance by x,m (millions)}
 \item {'N_yx'    - annual (July 1) abundance by x (millions)}
}


Uses \code{reshape2::melt} and \code{reshape2::dcast}.
}
