% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.RecSizeDistribution.R
\name{getMDFR.Pop.RecSizeDistribution}
\alias{getMDFR.Pop.RecSizeDistribution}
\title{Get recruitment size distributions from several model runs}
\usage{
getMDFR.Pop.RecSizeDistribution(obj, verbose = FALSE)
}
\arguments{
\item{obj}{- object with results for the models to be compared that can be converted to a list of tcsam2013.resLst objects}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format.
}
\description{
Function to get recruitment size distributions from 
several model runs.
}
\details{
Uses \code{reshape2} package.
}

