% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.PopProcesses.R
\name{getMDFR.PopProcesses}
\alias{getMDFR.PopProcesses}
\title{Get predicted population processes from several model runs}
\usage{
getMDFR.PopProcesses(reps, type = "R_cz")
}
\arguments{
\item{reps}{- list of objects derived from Jack's R files for the models to be compared}
}
\value{
dataframe
}
\description{
Function to get predicted population processes from 
several model runs.
}
\details{
Potential values for 'type' are:
\itemize{
 \item {'R_cz' - size distribution for recruitment}
 \item {'prM2M_cxz' - probability of molt-to-maturity}
 \item {'N_yxmsz' - numbers-at-size by x,m,s [TODO]}
}
}

