% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Surveys.Abundance.R
\name{getMDFR.Surveys.Abundance}
\alias{getMDFR.Surveys.Abundance}
\title{Get survey abundance time series from model results from TCSAM2013 model runs as a dataframe}
\usage{
getMDFR.Surveys.Abundance(obj, category = "index", cast = "y+x",
  fleets = "NMFS trawl survey", verbose = FALSE)
}
\arguments{
\item{obj}{- single tcsam2013.rep object, tcsam2013.resLst object, or named list of the latter}

\item{category}{- 'index' is only choice}

\item{cast}{- casting formula for excluding y,x,m,s,z factor levels from an average-at-size across unspecified factors}

\item{fleets}{- name(s) of surveys}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get survey abundance time series from model results from TCSAM2013 model runs as a dataframe.
}
\details{
Extracts the estimated survey abundance time series.
}

