% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Surveys.Biomass.R
\name{getMDFR.Surveys.Biomass}
\alias{getMDFR.Surveys.Biomass}
\title{Get survey biomass time series from model results from TCSAM2013 model runs as a dataframe}
\usage{
getMDFR.Surveys.Biomass(obj, category = "index", cast = "y+x",
  fleets = "NMFS trawl survey", verbose = FALSE)
}
\arguments{
\item{obj}{- single tcsam2013.rep object, tcsam2013.resLst object, or named list of the latter}

\item{category}{- 'index' is only choice}

\item{cast}{- casting formula (e.g., "x+m") for excluding x,m,s,z factor levels from a sum across the unspecified factors}

\item{fleets}{- name(s) of survey(s)}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get survey biomass time series from model results from TCSAM2013 model runs as a dataframe.
}
\details{
Extracts the estimated survey biomass time series.
}
