% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.SurveyTimeSeries.R
\name{getMDFR.SurveyTimeSeries}
\alias{getMDFR.SurveyTimeSeries}
\title{Get estimated/predicted surveys-related time series from several model runs}
\usage{
getMDFR.SurveyTimeSeries(reps, type = c("MB_xy", "B_xy", "N_yxmsz", "N_yxmz",
  "N_yxms", "N_yxm", "N_yx"), verbose = FALSE)
}
\arguments{
\item{reps}{- list of objects derived from Jack's R files for the models to be compared}

\item{type}{- survey-related time series to retrieve}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe
}
\description{
Function to get estimated/predicted surveys-related time series fom 
several model runs.
}
\details{
Potential values for 'type' are:
\itemize{
 \item {'MB_xy' - mature biomass at survey time, by sex (1000's t)}
 \item {'B_xy' - total biomass at survey time, by sex (1000's t)}
 \item {'N_yxms'  - annual abundance by x,m,s (millions)}
 \item {'N_yxm'   - annual abundance by x,m (millions)}
 \item {'N_yx'    - annual abundance by x (millions)}
 \item {'lglN_y'  - legal male annual abundance (millions)}
 \item {'lglB_y'  - legal male annual abundance (millions)}
 \item {'Pr_yxmsz'  - annual proportions-at-size by x,m,s}
 \item {'selSrv_cxz' - survey selectivity, by time period and sex}
}
}

