% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.TimeSeries1.R
\name{getMDFR.TimeSeries1}
\alias{getMDFR.TimeSeries1}
\title{Get estimated/predicted time series from several model runs}
\usage{
getMDFR.TimeSeries1(reps, type = "MB_xy")
}
\arguments{
\item{reps}{- list of objects derived from Jack's R files for the models to be compared}

\item{type}{- time series to retrieve}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe
}
\description{
Function to get estimated/predicted time series fom 
several model runs.
}
\details{
Potential values for 'type' are:
\itemize{
 \item {'MB_xy' - mature biomass at mating, bysex (1000's t)}
 \item {'rec_y' - total recruitment (millions)}
}
}

