% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNatMort.R
\name{getNatMort}
\alias{getNatMort}
\title{Function to get natural mortality rates by year from an .std file or dataframe}
\usage{
getNatMort(sdobj, styr = NULL, endyr = NULL)
}
\arguments{
\item{sdobj}{- object (dataframe) obtained from reading the TCSAM2013 .std file}

\item{styr}{- model start year}

\item{endyr}{- model end year}
}
\value{
dataframe with columns sex,shell,maturity,year,val,std
}
\description{
This function get natural mortality estimates by year
  and sex for immatures, new shell matures and old shell matures.
}

