% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOFC.R
\name{getOFC}
\alias{getOFC}
\title{Create a tcsam2013.ofc object from a TCSAM2013.final_likelihood_components.csv file}
\usage{
getOFC(inp.dir = "./")
}
\arguments{
\item{inp.dir}{= folder from which to read TCSAM2013.final_likelihood_components.csv file}
}
\value{
dataframe object corresponding to the TCSAM2013.final_likelihood_components.csv file, or NULL if file does not exist. 
Returned object has class 'tcsam2013.ofc'.
}
\description{
Function to create a tcsam2013.ofc object from a TCSAM2013.final_likelihood_components.csv file.
}
\details{
Uses \code{wtsUtilities::selectFile} to open a file dialog if inp.dir is NULL.
}

