% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPar.R
\name{getPar}
\alias{getPar}
\title{Create a TCSAM2013 par object from a par file.}
\usage{
getPar(in.par = NULL)
}
\arguments{
\item{in.par}{= filename of par file}
}
\value{
list object corresponding to the par file (or NULL if file does not exist)
}
\description{
Function to create a TCSAM2013 par object from a par file.
}
\details{
Uses \code{wtsUtilities::selectFile} to open a file dialog if in.par is NULL.
}

