% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPRS.R
\name{getPRS}
\alias{getPRS}
\title{Create a TCSAM2013 prs object from a params csv file.}
\usage{
getPRS(inp = NULL)
}
\arguments{
\item{inp}{- filename for params csv file (or NULL)}
}
\value{
list object corresponding to the params file, or NULL if file does not exist
}
\description{
Function to create a TCSAM2013 prs object from a params csv file.
}
\details{
Uses \code{wtsUtilities::selectFile()}.
}

