% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRep.R
\name{getRep}
\alias{getRep}
\title{Create a TCSAM2013 rep object from a TCSAM_OLDSTYLE R file}
\usage{
getRep(inp = NULL)
}
\arguments{
\item{inp}{= filename of TCSAM_OLDSTYLE R file}
}
\value{
list object corresponding to the TCSAM_OLDSTYLE R file, or NULL if file does not exist. 
Returned object has class 'tcsam213.rep'.
}
\description{
Function to create a TCSAM2013 rep object from a TCSAM_OLDSTYLE R file.
}
\details{
Uses \code{wtsUtilities::selectFile} to open a file dialog if in.par is NULL.
Uses \code{PBSmodelling::readList} to read file.
}

