% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResLst.R
\name{getResLst}
\alias{getResLst}
\title{Create a tcsam2013.resLst object from a model run}
\usage{
getResLst(inp.dir = NULL, model = "tcsam2013alta",
  rep = "TCSAM2013.OLDSTYLE.final.R")
}
\arguments{
\item{inp.dir}{- folder with model output}

\item{model}{- name of model executable}

\item{rep}{- filename to read for rep object}
}
\value{
a tcsam2013.resLst object.
}
\description{
Function to create a tcsam2013.resLst object from a model run.
}
\details{
Uses \code{wtsUtilities::selectFile} to open a file dialog to select model directory 
if inp.dir is NULL. A tcsam2013.resLst object is a list with elements
\itemize{
 \item{rep - a tcsam2013.rep object}
 \item{prs - a tcsam2013.prs object}
 \item{std - a tcsam2013.std object, or NULL}
 \item{ofc - a tcsam2013.ofc object}
}
}

