% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRunCommands.R
\name{getRunCommands}
\alias{getRunCommands}
\title{Generate run commands for a tcsam2013 model run}
\usage{
getRunCommands(os = "osx", model = "tcsam2013alta", path2model = "",
  configFile = "", pin = FALSE, hess = FALSE, mcmc = FALSE,
  jitter = FALSE, seed = NULL)
}
\arguments{
\item{os}{- 'win' or 'mac'}

\item{model}{-}

\item{path2model}{-}

\item{pin}{-}

\item{hess}{-}

\item{mcmc}{-}

\item{jitter}{-}

\item{seed}{-}
}
\description{
Function to generate a script to make a tcsam2013 model run
}
\details{

}

