% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStd.R
\name{getStd}
\alias{getStd}
\title{Create a dataframe (an std object) from an ADMB std file}
\usage{
getStd(inp = NULL)
}
\arguments{
\item{inp}{- name of the ADMB std file}
}
\value{
dataframe of class 'tcsam2013.std', corresponding to the std file, or NULL if file does not exist
}
\description{
Function to create a dataframe (an std object) from an ADMB std file.
}
\details{
Uses \code{wtsUtilities::selectFile} to open a file dialog if in.par is NULL.
}
