% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubPaths.R
\name{getSubPaths}
\alias{getSubPaths}
\title{Extract subpaths from a vector of file paths}
\usage{
getSubPaths(paths, last = NULL, first = NULL)
}
\arguments{
\item{paths}{- vector of paths to extract from}

\item{last}{- number of final folder levels to retain}

\item{first}{- number of initial folder levels to retain}
}
\value{
character vector of subpaths
}
\description{
Function to extract subpaths from a vector of file paths.
}
\details{
If first and last are both given, only the "last" subpaths will be exported.
}
