% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeInfo.R
\name{getTimeInfo}
\alias{getTimeInfo}
\title{Get time information for a number of model runs}
\usage{
getTimeInfo(obj)
}
\arguments{
\item{obj}{- object to convert (see details)}
}
\value{
List with model time information
}
\description{
Function to get time information for a number of model runs.
}
\details{
'obj' can be one of the following:
\itemize{
 \item{a list of tcsam2013.resLst objects}
 \item{a single tcsam2013.resLst object}
 \item{a list of tcsam2013.rep objects}
 \item{a single tcsam2013.rep object}
}
}
