% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWTS.R
\name{getWTS}
\alias{getWTS}
\title{Create a TCSAM2013 wts object from a TCSAM_WTS R file}
\usage{
getWTS(inp = NULL)
}
\arguments{
\item{inp}{- filename of TCSAM_WTS R file (or NULL)}
}
\value{
list object corresponding to the file, or NULL if file does not exist
}
\description{
Function to create a TCSAM2013 wts object from a TCSAM_WTS R file.
}
\details{
Uses \code{wtsUtilities::selectFile()}.
}
