% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotABC.R
\name{plotABC}
\alias{plotABC}
\title{Function to plot ABC and OFL calculations from calcABC(...)}
\usage{
plotABC(x, OFL, ABC.pstar, ABC.buff, buffer = 0.2, title = NULL,
  xlims = NULL, save = FALSE)
}
\arguments{
\item{x}{- empirical cdf of OFL}

\item{OFL}{- estimated OFL}

\item{ABC.pstar}{- ABC based on pstar method}

\item{ABC.buff}{- ABC based on buffer method}

\item{buffer}{- buffer used for ABC.buff}

\item{title}{- title for plot}

\item{xlims}{- range for x-axis in plot}

\item{save}{- flag (T/F) to save OFL plot to "oflPlot.png".}
}
\description{
This function plots ABC and OFL from TCSAM2013 projection model output.
}
\details{
None.
}

