% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.Fishery.Discarded.NatZ.R
\name{plotData.Fishery.Discarded.NatZ}
\alias{plotData.Fishery.Discarded.NatZ}
\title{Function to plot numbers-at-size from fishery discard data.}
\usage{
plotData.Fishery.Discarded.NatZ(nAtZ, label = "")
}
\arguments{
\item{nAtZ}{- discarded catch numbers-at size 3d array (sex,shell condition, year, size).}

\item{label}{- label for plot}
}
\description{
This function plots discarded catch numbers-at-size data by year as circles.
New shell crab are plotted in blue, old shells in green. Plots are created by sex.
}
\details{
Uses \code{wtsPlots::plotCompsAsCircles}.
}

