% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.Fishery.Retained.NatZ.R
\name{plotData.Fishery.Retained.NatZ}
\alias{plotData.Fishery.Retained.NatZ}
\title{Function to plot numbers-at-size from retained fishery data.}
\usage{
plotData.Fishery.Retained.NatZ(nAtZ,
  label = "directed fishery: retained males")
}
\arguments{
\item{nAtZ}{- retained catch numbers-at size 3d array (shell condition, year, size).}

\item{label}{- label for plot}
}
\description{
This function plots retained catch numbers-at-size data by year as circles.
New shell crab are plotted in blue, old shells in green.
}
\details{
Uses \code{wtsPlots::plotCompsAsCircles}.
}

