% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotModelComparisonsGG.TimeSeries.R
\name{plotModelComparisonsGG.TimeSeries}
\alias{plotModelComparisonsGG.TimeSeries}
\title{Plot comparisons of time series from a set of model runs}
\usage{
plotModelComparisonsGG.TimeSeries(dfr, position = "dodge", facets = NULL,
  scales = "fixed", plotObs = TRUE, plotMod = TRUE, ci = 0.95,
  pdfType = "lognormal", xlab = "year", ylab = NULL, title = NULL,
  xlims = NULL, ylims = NULL, showPlot = TRUE)
}
\arguments{
\item{dfr}{- dataframe}

\item{position}{- indicates ggplot2 position_ to use ('dodge','jitter','identity',)}

\item{facets}{- string giving faceting formula}

\item{scales}{- ggplot2 scales option for facet_grid}

\item{plotObs}{- plot observations}

\item{plotMod}{- plot model fits/predictions}

\item{ci}{- confidence interval for error bars}

\item{pdfType}{- assumed error distribution for confidence intervals}

\item{xlab}{-}

\item{ylab}{-}

\item{title}{-}

\item{xlims}{-}

\item{ylims}{-}

\item{showPlot}{-}
}
\value{
ggplot object
}
\description{
Function to plot a comparison of time series from a set of model runs.
}
\details{

}

