% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuadPlot.R
\name{plotQuadPlot}
\alias{plotQuadPlot}
\title{Plot a quad plot from a TCSAM report object}
\usage{
plotQuadPlot(obj.rep, F35 = NA, B35 = NA)
}
\arguments{
\item{obj.rep}{- report file object}

\item{F35}{- F35 from a projection model run}

\item{B35}{- B35 from a projection model run}
}
\value{
nothing
}
\description{
Function to plot a quad plot from a TCSAM report object.
}

