% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTCSAM_WTS.R
\name{plotTCSAM_WTS}
\alias{plotTCSAM_WTS}
\title{Plot TCSAM_WTS...R output from a TCSAM2013 model run}
\usage{
plotTCSAM_WTS(res = NULL, zscores.facet.scales = "fixed",
  showPlot = FALSE)
}
\arguments{
\item{res}{- TCSAM_WTS-type list}

\item{zscores.facet.scales}{- flag to control zscore scales for bycatch fisheries ('fixed','free_y')}

\item{showPlot}{- flag (T/F) to show plots immediately}
}
\value{
list of lists with ggplot2 objects as elements:
\itemize{
 \item zscores - srv, fsh plots
}
}
\description{
Function plot TCSAM_WTS...R output from a TCSAM2013 model run.
}
\details{
None.
}
