% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTCSAM_WTS.ZScores.R
\name{plotTCSAM_WTS.ZScores}
\alias{plotTCSAM_WTS.ZScores}
\title{Plot z-scores from a TCSAM2013 model run}
\usage{
plotTCSAM_WTS.ZScores(res = NULL, facet.scales = "fixed",
  showPlot = FALSE)
}
\arguments{
\item{res}{- TCSAM_WTS-type list}

\item{facet.scales}{- flag to control scales for bycatch fisheries ('fixed','free_y')}

\item{showPlot}{- flag (T/F) to show plots immediately}
}
\value{
list with ggplot2 objects as elements:
\itemize{
 \item srv - plot for survey
 \item fsh - plot for fisheries
}
}
\description{
Function to plot z-scores from a TCSAM2013 model run.
}
\details{
Uses \code{reshape2::melt()}.
}

