% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZScores.R
\name{plotZScores}
\alias{plotZScores}
\title{Plot z-scores from a set of model runs}
\usage{
plotZScores(dfr, x = "y", y = "zscore", color = NULL, shape = NULL,
  position = "identity", dodge = 0.2, facets = NULL,
  facet.scales = "fixed", xlab = "year", ylab = NULL, title = NULL,
  legend = NULL, xlims = NULL, ylims = NULL, showPlot = TRUE)
}
\arguments{
\item{dfr}{- dataframe}

\item{x}{- column name for x axis (default = 'year')}

\item{y}{- column name for y axis (default='z-score')}

\item{color}{- column name for color levels (or NULL)}

\item{shape}{- column name for shape levels (or NULL)}

\item{position}{- indicates ggplot2 position_ to use ('dodge','jitter','identity',)}

\item{facets}{- string giving faceting formula for facet_grid}

\item{facet.scales}{- ggplot2 scales option for facet_grid}

\item{xlab}{- label for x axis}

\item{ylab}{- label for y axis}

\item{title}{- title for plot}

\item{xlims}{- limits for x axis}

\item{ylims}{- limits for y axis}

\item{showPlot}{- flag (T/F) to show plot immediately}
}
\value{
ggplot2 object
}
\description{
Function to plot z-scores from a set of model runs.
}
\details{
None.
}

