% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printGGList.R
\name{printGGList}
\alias{printGGList}
\title{Print a ggplot2 object, or list of ggplot2 objects}
\usage{
printGGList(plots, figno = NULL, cap = NULL)
}
\arguments{
\item{plots}{- a ggplot2 object or a list of ggplot2 objects}

\item{figno}{- figure number (or NULL)}

\item{cap}{- plot caption}
}
\value{
figure number for next plot, if figno was not NULL.
}
\description{
Function to print a ggplot2 object or a list of ggplot2 objects.
}
\details{
Captions for each figure are taken from the name of the list element associated with the plot.
\code{figno} is substituted for '&&fno' in the caption and incremented for the next plot.
}

