% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderReport.R
\name{renderReport}
\alias{renderReport}
\title{Render report}
\usage{
renderReport(paths = NULL, out.dir = getwd())
}
\arguments{
\item{paths}{- list of paths to model runs}
}
\value{
List with paths to model results as elements.
}
\description{
renders report.
}
\details{
Renders inst/Rmd/plotTCSAM2013.Rmd in rTCAM2013 package. Requires knittr, rmarkdown.
}

