% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderReports.R
\name{renderReports}
\alias{renderReports}
\title{Render reports}
\usage{
renderReports(paths = NULL, obj = NULL, numRecent = 15,
  plot1stObs = TRUE, out.dir = getwd(), figures = TRUE, tables = TRUE,
  format = "word_document", clean = FALSE)
}
\arguments{
\item{paths}{- vector of paths to model runs, with case names as names (optional)}

\item{obj}{- object convertible to a list of tcsam2013.resLst objects, with case names as names (optional)}

\item{out.dir}{- folder for output}

\item{figures}{- flag (T/F) to create}

\item{tables}{- flag (T/F) to create}

\item{format}{- output file format ("word_document","pdf_document", or "html_document")}

\item{clean}{- flag (T/F) to create}
}
\value{
vector with paths to model results as elements, if paths is not NULL.
}
\description{
renders reports.
}
\details{
Renders createReport.Tables.Rmd and createReport.Figures.Rmd from 
the inst/Rmd folder in the rTCAM2013 package. 
Values for 'paths' and 'obj' should not be given simultaneously. 
Output files will be TCSAM2013ReportFigures.ext and TCSAM2013ReportTables.ext,
where 'ext' is 'pdf' or 'docx', depending on the format chosen.
}

