% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runProfiles.R
\name{runProfiles}
\alias{runProfiles}
\title{Function to run profiles on a TCSAM2013 model.}
\usage{
runProfiles(os = "osx", path = ifelse(tolower(os) == "win", ".\\\\", "./"),
  model = "tcsam2013alta", path2model = "", configFile = NULL,
  controlFile = NULL, profiles = NULL, runSeqs = TRUE, numRuns = 4)
}
\arguments{
\item{os}{- 'win' or 'mac' or 'osx'}

\item{path}{- path to parent folder for model output}

\item{model}{- TCSAM2013 model executable name}

\item{path2model}{- path to model executable}

\item{configFile}{- path to model configuration file template}

\item{controlFile}{- path to model control file template}

\item{profiles}{- list of lists of variables to profile on}

\item{runSeqs}{- TRUE to run sequences, FALSE to jitter}

\item{numRuns}{- number of sequences or jitters to run}
}
\value{
- par file dataframe
}
\description{
This functions runs profiles on a TCSAM2013 model.
}
\details{


This function creates a shell script ('./tmp.sh') in the
working directory and uses it to run a version of the TCSAM2013 model.
Initial model parameters can be jittered based on the system clock time as a seed
to the random number generator. The seed and final objective function value are
saved for each model run in a csv file (the value of out.csv).
}

