% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runTCSAM2013.R
\name{runTCSAM2013}
\alias{runTCSAM2013}
\title{Function to run TCSAM2013 .}
\usage{
runTCSAM2013(path = "./", os = "osx", model = "tcsam2013alta",
  path2model = "", configFile = "", pin = FALSE, hess = FALSE,
  mcmc = FALSE, jitter = FALSE, seed = NULL)
}
\arguments{
\item{path}{- path for model output}

\item{os}{- 'win' or 'mac' or 'osx'}

\item{model}{- TCSAM2013 model executable name}

\item{path2model}{- path to model executable}

\item{pin}{- T/F to use a pin file}

\item{hess}{- T/F to compute hessian (and .std file)}

\item{mcmc}{- T/F to run mcmc}

\item{jitter}{- T/F to jitter parameters}

\item{seed}{- seed for random number generator (or NULL)}

\item{in.csv}{- filename for jitter info (seed, obj fun value) from ADMB model run}

\item{out.csv}{- filename for jittered results}
}
\value{
- par file dataframe
}
\description{
This functions runs a TCSAM2013 model once.
}
\details{


This function creates a shell script ('./tmp.sh') in the
working directory and uses it to run a version of the TCSAM2013 model.
Initial model parameters can be jittered based on the system clock time as a seed
to the random number generator. The seed and final objective function value are
saved for each model run in a csv file (the value of out.csv).
}

