% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcABC.R
\name{calcABC}
\alias{calcABC}
\title{Function to calculate ABC and OFL from projection model output}
\usage{
calcABC(res = NULL, fn = NULL, pstar = 0.49, buffer = 0.1,
  xlims = NULL, title = NULL)
}
\arguments{
\item{res}{- r list from TCSAM2013 projection model run}

\item{fn}{- filename of TCSAM2013 projection model run r file}

\item{pstar}{-}

\item{buffer}{- buffer for ABC as fraction (i.e., 0.20 = 20 \% yields ABC = 0.8*OFL)}

\item{xlims}{- range for x-axis in plot}

\item{title}{- title for plot}
}
\description{
This function calculates ABC and OFL from TCSAM2013 projection model output.
}
\details{
Uses standard graphics. Creates "OFL_plot.png".
}

