% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkParams.R
\name{checkParams}
\alias{checkParams}
\title{Function to plot parameter values.}
\usage{
checkParams(obj.prs = NULL, obj.std = NULL, dp = 0.01)
}
\arguments{
\item{obj.prs}{- dataframe or filename for csv file containing parameters information}

\item{obj.std}{- dataframe obtained from reading, or filename of, the .std file}

\item{dp}{- percent difference between parameter value and upper/lower limits used to color plot}
}
\value{
- list
}
\description{
This function plots parameters values, together with their limits
(if any) and the posterior distributions implied by their estimated standard
errors.
}
\details{

}

