% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.FisheryPRs.R
\name{compareModelResults.FisheryPRs}
\alias{compareModelResults.FisheryPRs}
\title{Compare Pearson's residuals from fishery size comps among several model runs}
\usage{
compareModelResults.FisheryPRs(obj, showPlot = FALSE, pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- name for output pdf file}
}
\value{
list of ggplot2 objects
}
\description{
Function to Pearson's residuals from fishery size comps among 
several model runs.
}
\details{
Uses \code{getMDFR.FisheryQuantities}.
}
