% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareModelResults.NatMort.R
\name{compareModelResults.NatMort}
\alias{compareModelResults.NatMort}
\title{Function to plot natural mortality rates by year using ggplot2}
\usage{
compareModelResults.NatMort(resLst, ci = 0.8, min.yr = NULL,
  max.yr = NULL, showPlot = TRUE)
}
\arguments{
\item{resLst}{- list of results objects}

\item{ci}{- confidence interval for error bars}

\item{min.yr}{- min year to plot}

\item{max.yr}{- max year to plot}
}
\value{
list with dataframe and list of plots as elements
}
\description{
This function plots natural mortality estimates by year
  and sex for immatures, new shell matures and old shell matures.
}
\details{

}

