% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.Params.R
\name{compareModelResults.Params}
\alias{compareModelResults.Params}
\title{Function to compare parameter values from different models.}
\usage{
compareModelResults.Params(objs, dp = 0.01, fac = 3, nc = 3, nr = 4,
  showPlot = TRUE, pdf = "ModelComparisons.Params.pdf")
}
\arguments{
\item{objs}{- model results objects (each is a list with elements 'prs' and 'std')}

\item{dp}{- percent difference between parameter value and upper/lower limits used to flag outliers}

\item{fac}{- number of std devs to extend uncertainty plots}

\item{nc}{- number of columns of plots per page}

\item{nr}{- number of rows of plots per page}

\item{showPlot}{- flag to show plots}

\item{pdf}{- file name for printing plots to a pdf file (or NULL to print to screen)}
}
\value{
- list with dfr, vfr, and plots as elements
}
\description{
This function extracts and plots parameters values, together with their limits
(if any) and the posterior distributions implied by their estimated standard
errors from several models.
}
\details{
Uses \code{reshape2::melt()}.
}

