% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.Params.R
\name{compareModelResults.Params}
\alias{compareModelResults.Params}
\title{Function to compare parameter values from different models}
\usage{
compareModelResults.Params(obj, dp = 0.01, showPlot = FALSE, pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{dp}{- percent difference between parameter value and upper/lower limits used to flag outliers}

\item{showPlot}{- flag to show plots}

\item{pdf}{- file name for printing plots to a pdf file (or NULL to print to screen)}
}
\value{
- list of lists with ggplot2 objects as elements
}
\description{
This function extracts and plots parameters values from several models, 
together with their limits(if any) and their estimated standard errors (if any).
}
\details{
Uses \code{getMDFR.ParamsPlusStdDevs()}. Returned list has top level elements:
\itemize{
 \item{population}
 \item{surveys}
 \item{fisheries}
}
}
