% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.PopProcesses.R
\name{compareModelResults.PopProcesses}
\alias{compareModelResults.PopProcesses}
\title{Compare predicted population processes among several model runs}
\usage{
compareModelResults.PopProcesses(obj, showPlot = FALSE, pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- name for output pdf file}
}
\value{
list of ggplot2 objects, returned invisibly.
}
\description{
Function to compare predicted population processes among 
several model runs.
}
\details{
None
}
