% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.PopProcesses.R
\name{compareModelResults.PopProcesses}
\alias{compareModelResults.PopProcesses}
\title{Compare estimated/predicted time series among several model runs}
\usage{
compareModelResults.PopProcesses(reps = NULL, cases = NULL,
  showPlot = FALSE, pdf = "ModelComparison.PopProcesses.pdf")
}
\arguments{
\item{reps}{- list of objects derived from Jack's R files for the models to be compared}

\item{cases}{- vector of labels for model cases (if 'reps' is not given)}

\item{pdf}{- name for output pdf file}

\item{numRecent}{- number of recent years to plot}

\item{styr}{- start year for model run}

\item{endyr}{- final year of model run (assessment year)}

\item{plot1stObs}{- flag to plot observations from the first case, only}

\item{obsyr}{- start year for survey observations}

\item{pltyr}{- start year for plots with only model-predicted values}

\item{reclag}{- recruitment lag}
}
\value{
vector of list objects corresponding to the objects returned by each model R file.
}
\description{
Function to compare estimated/predicted time series among 
several model runs.
}
\details{
If 'reps' is not given, then the user is prompted to select Jack's R file output from each 
model to be compared. If 'cases' is given, the user is prompted to select the file
corresponding to each case. If 'cases' is not given, then the user may select an 
arbitrary number of files (one at a time), ending selection by pressing 'cancel' on the selection box'.\cr\cr
If 'reps' is not given, the working directory is set two levels above the 1st model case file selected.\cr\cr
Uses \code{PBSmodelling::readList} and \code{wtsUtilities::selectFile}.
}

