% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.PrM2M.R
\name{compareModelResults.PrM2M}
\alias{compareModelResults.PrM2M}
\title{Compare probability of molt-to-maturity among several model runs}
\usage{
compareModelResults.PrM2M(obj, showPlot = FALSE, pdf = NULL)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{showPlot}{- flag to print plot to current device}

\item{pdf}{- name for output pdf file}
}
\value{
ggplot object
}
\description{
Function to compare probability of molt-to-maturity among several model runs.
}
\details{
None.
}

