% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelResults.SurveyQuantities.R
\name{compareModelResults.SurveyQuantities}
\alias{compareModelResults.SurveyQuantities}
\title{Compare estimated/predicted survey quatities among several model runs}
\usage{
compareModelResults.SurveyQuantities(reps = NULL, numRecent = 15,
  plot1stObs = TRUE, showPlot = FALSE,
  pdf = "ModelComparisons.SurveyTimeSeries.pdf")
}
\arguments{
\item{reps}{- list of objects derived from Jack's R files for the models to be compared}

\item{numRecent}{- number of recent years to plot}

\item{plot1stObs}{- flag to plot observations from the first case, only}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- name for output pdf file}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare estimated/predicted survey quantities among 
several model runs.
}
\details{
Uses \code{getMDFR.SurveyQuantities}.
}

