% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.FisheryQuantities.R
\name{getMDFR.FisheryQuantities}
\alias{getMDFR.FisheryQuantities}
\title{Get estimated/predicted fishery-related quantities from several model runs}
\usage{
getMDFR.FisheryQuantities(obj, type = c("bio.retm", "bio.dscm", "bio.totm",
  "prNatZ.ret", "prNatZ.tot", "PRs.ret", "PRs.tot", "mnPrNatZ.ret",
  "mnPrNatZ.tot", "sel_cxz", "sel_yxz", "ret_yxz", "zscores.tot", "zscores.ret",
  "effSS.ret", "effSS.tot", "qFsh_xy", "maxFc_xy", "max rates", "mean rates"),
  ci = 0.8, pdfType = c("norm2", "normal", "lognormal"), verbose = FALSE)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{type}{- fishery-related quantity to retrieve}

\item{ci}{- confidence interval for error bars on observations}

\item{pdfType}{- type of error distribution for observations}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get estimated/predicted fishery-related quantities from 
several model runs.
}
\details{
Potential values for 'type' are:
\itemize{
 \item {'bio.retm' - retained catch mortality (1000's t)}
 \item {'bio.dscm' - discard catch mortality (1000's t) for males in directed fishery}
 \item {'bio.totm' - total catch mortality (1000's t)}
 \item {'prNatZ.ret' - annual retained catch proportions-at-size}
 \item {'prNatZ.tot' - annual total catch proportions-at-size}
 \item {'PRs.ret'      - pearsons residuals for retained catch proportions-at-size}
 \item {'PRs.tot'      - pearsons residuals for total catch proportions-at-size}
 \item {'mnPrNatZ.ret' - mean proportions-at-size for retained catch data}
 \item {'mnPrNatZ.tot' - mean proportions-at-size for total catch data}
 \item {'sel_cxz' - fishery selectivity and retention functions, by time period and sex}
 \item {'sel_yxz' - fishery selectivity functions, by year and sex}
 \item {'ret_yxz' - fishery retention functions, by year and sex}
 \item {'zscores.tot' - annual z-scores for fits to total catch biomass}
 \item {'zscores.ret' - annual z-scores for fits to retained catch biomass}
 \item {'effSS.tot' - effective (and input) sample sizes for total catch size comps}
 \item {'effSS.ret' - effective (and input) sample sizes for retained catch size comps}
 \item {'qFsh_xy' - annual fishery catchabilities}
 \item {'maxFc_xy' - annual max capture rates}
 \item {'max rates' - max fishing mortality, retained mortality, and capture rates}
 \item {'mean rates' - mean fishing mortality, retained mortality, and capture rates}
}
Requires sqldf package.
}

