% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.PopProcesses.R
\name{getMDFR.PopProcesses}
\alias{getMDFR.PopProcesses}
\title{Get predicted population processes from several model runs}
\usage{
getMDFR.PopProcesses(reps, type = c("M_yxm", "R_cz", "prM2M_cxz", "mnZAM_cxz",
  "T_cxzz"), verbose = FALSE)
}
\arguments{
\item{reps}{- tcsam2013.rep object (derived from OLDSTYLE R files) for the models to be compared}

\item{type}{- population process to retrieve}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe with columns 'modeltype','model',
'y' or 'pc','x','m' or 'z', and 'val' (or 'zp' and 'val').
}
\description{
Function to get predicted population processes from 
several model runs.
}
\details{
Potential values for 'type' are:
\itemize{
 \item {'M_yxm' - natural mortality rates}
 \item {'R_cz' - size distribution for recruitment}
 \item {'prM2M_cxz' - probability of molt-to-maturity}
 \item {'mnZAM_cxz' - mean growth increment}
 \item {'T_cxzz' - growth transition matrix}
}
Uses \code{reshape2::melt}.
}

