% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.PopProcesses.R
\name{getMDFR.PopProcesses}
\alias{getMDFR.PopProcesses}
\title{Get predicted population processes from several model runs}
\usage{
getMDFR.PopProcesses(obj, type = c("M_yxm", "R_cz", "prM2M_cxz", "mnZAM_cxz",
  "T_cxzz"), verbose = FALSE)
}
\arguments{
\item{obj}{- object with results for the models to be compared that can be converted to a list of tcsam2013.resLst objects}

\item{type}{- population process to retrieve}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format.
}
\description{
Function to get predicted population processes from 
several model runs.
}
\details{
Potential values for 'type' are:
\itemize{
 \item {'M_yxm' - natural mortality rates}
 \item {'R_cz' - size distribution for recruitment}
 \item {'prM2M_cxz' - probability of molt-to-maturity}
 \item {'mnZAM_cxz' - mean growth increment}
 \item {'T_cxzz' - growth transition matrix}
}
Uses \code{reshape2::melt}.
}

