% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.PopQuants.R, R/getMDFR.TimeSeries1.R
\name{getMDFR.PopQuants}
\alias{getMDFR.PopQuants}
\title{Get predicted population quantities (time series) from several model runs}
\usage{
getMDFR.PopQuants(reps, type = "MB_xy")

getMDFR.PopQuants(reps, type = "MB_xy")
}
\arguments{
\item{reps}{- list of objects derived from Jack's R files for the models to be compared}

\item{reps}{- list of objects derived from Jack's R files for the models to be compared}
}
\value{
dataframe

dataframe
}
\description{
Function to get predicted population quantities (time series) fom 
several model runs.

Function to get estimated/predicted time series fom 
several model runs.
}
\details{
Potential values for 'type' are:
\itemize{
 \item {'MB_yx' - mature biomass at mating, by sex (1000's t)}
 \item {'R_y' - total recruitment (millions)}
 \item {'N_yxmsz' - numbers-at-size by x,m,s [TODO]}
}

Potential values for 'type' are:
\itemize{
 \item {'MB_xy' - mature biomass at mating, bysex (1000's t)}
 \item {'rec_y' - total recruitment (millions)}
}
}

