% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.SurveyQuantities.R
\name{getMDFR.SurveyQuantities}
\alias{getMDFR.SurveyQuantities}
\title{Get estimated/predicted survey-related quantities from several model runs}
\usage{
getMDFR.SurveyQuantities(obj, type = c("MB_yx", "lglN_y", "lglB_y", "N_yxmsz",
  "N_yxmz", "N_yxz", "N_yxms", "N_yxm", "N_yx", "B_yxmsz", "B_yxmz", "B_yxz",
  "B_yxms", "B_yxm", "B_yx", "prNatZ_yxmsz", "prNatZ_yxmz", "prNatZ_yxz",
  "PRs_yxmsz", "PRs_yxmz", "PRs_yxz", "mnPrNatZ_xmz", "mnPrNatZ_xz", "qSrv_xy",
  "selSrv_cxz", "selSrv_yxz", "zscores", "effSS_y"), pdfType = c("lognormal",
  "normal"), ci = 0.8, verbose = FALSE)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{type}{- survey-related quantity to retrieve}

\item{pdfType}{- type of error distribution for observations}

\item{ci}{- confidence interval for error bars on observations}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe with columns in canonical format
}
\description{
Function to get estimated/predicted survey-related quantities from 
several model runs.
}
\details{
Potential values for 'type' are:
\itemize{
 \item {'MB_yx' - mature biomass at survey time, by sex (1000's t)}
 \item {'B_yx' - total biomass at survey time, by sex (1000's t)}
 \item {'N_yxmsz'  - annual survey abundance-at-size by x,m,s (millions)}
 \item {'N_yxmz'  - annual survey abundance-at-size by x,m (millions)}
 \item {'N_yxz'  - annual survey abundance-at-size by x (millions)}
 \item {'N_yxms'  - annual survey abundance by x,m,s (millions)}
 \item {'N_yxm'   - annual survey abundance by x,m (millions)}
 \item {'N_yx'    - annual survey abundance by x (millions)}
 \item {'B_yxmsz'  - annual survey biomass-at-size by x,m,s,z (1000's t)}
 \item {'B_yxms'  - annual survey biomass by x,m,s (1000's t)}
 \item {'B_yxm'   - annual survey biomass by x,m (1000's t)}
 \item {'B_yx'    - annual survey biomass by x (1000's t)}
 \item {'lglN_y'  - legal male annual survey abundance (millions)}
 \item {'lglB_y'  - legal male annual survey biomass (1000's t)}
 \item {'prNatZ_yxmsz' - annual proportions-at-size by x,m,s}
 \item {'prNatZ_yxmz'  - annual proportions-at-size by x,m}
 \item {'prNatZ_yxz'   - annual proportions-at-size by x}
 \item {'PRs_yxmsz'    - pearsons residuals for annual proportions-at-size by x,m,s}
 \item {'PRs_yxmz'     - pearsons residuals for annual proportions-at-size by x,m}
 \item {'PRs_yxz'      - pearsons residuals for annual proportions-at-size by x}
 \item {'mnPrNatZ_xmz' - mean proportions-at-size by x,m}
 \item {'mnPrNatZ_xz'  - mean proportions-at-size by x}
 \item {'qSrv_yxz'   - survey catchability, by survey, year and sex}
 \item {'selSrv_cxz' - survey selectivity, by time period and sex}
 \item {'selSrv_yxz' - survey selectivity, by survey, year and sex}
 \item {'zscores'  - annual z-scores for fit to mature survey biomass, by sex}
 \item {'effSS_y' - effective (and input) sample sizes for multinomial fits}
}
Requires sqldf package.
}

