% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.SurveyQuantities.R
\name{getMDFR.SurveyQuantities}
\alias{getMDFR.SurveyQuantities}
\title{Get estimated/predicted surveys-related quantities from several model runs}
\usage{
getMDFR.SurveyQuantities(reps, type = c("MB_yx", "B_yx", "lglN_y", "lblB_y",
  "N_yxms", "N_yxm", "N_yx", "N_yxmsz", "N_yxmz", "N_yxz", "prNatZ_yxmsz",
  "prNatZ_yxmz", "prNatZ_yxz", "PRs_yxz", "mnPrNatZ_xz", "zscrs_yx", "effSS_y"),
  verbose = FALSE)
}
\arguments{
\item{reps}{- list of objects derived from Jack's R files for the models to be compared}

\item{type}{- survey-related time series to retrieve}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe
}
\description{
Function to get estimated/predicted surveys-related quantities fom 
several model runs.
}
\details{
Potential values for 'type' are:
\itemize{
 \item {'MB_yx' - mature biomass at survey time, by sex (1000's t)}
 \item {'B_yx' - total biomass at survey time, by sex (1000's t)}
 \item {'N_yxms'  - annual abundance by x,m,s (millions)}
 \item {'N_yxm'   - annual abundance by x,m (millions)}
 \item {'N_yx'    - annual abundance by x (millions)}
 \item {'lglN_y'  - legal male annual abundance (millions)}
 \item {'lglB_y'  - legal male annual abundance (millions)}
 \item {'NatZ_yxmsz' - annual abundance-at-size by x,m,s (millions)}
 \item {'NatZ_yxmz'  - annual abundance-at-size by x,m (millions)}
 \item {'NatZ_yxz'   - annual abundance-at-size by x (millions)}
 \item {'prNatZ_yxmsz' - annual proportions-at-size by x,m,s}
 \item {'prNatZ_yxmz'  - annual proportions-at-size by x,m}
 \item {'prNatZ_yxz'   - annual proportions-at-size by x}
 \item {'PRs_yxz'      - pearsons residuals for annual proportions-at-size by x}
 \item {'mnPrNatZ_xz'  - mean proportions-at-size by x}
 \item {'selSrv_cxz' - survey selectivity, by time period and sex}
 \item {'zscrs_yx'  - annual z-scores for fit to mature survey biomass, by sex}
 \item {'effSS_y' - effective (and input) sample sizes for multinomial fits}
}
Requires sqldf package.
}

