% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPar.R
\name{getPar}
\alias{getPar}
\title{Create a dataframe from an ADMB par file.}
\usage{
getPar(inp = NULL)
}
\arguments{
\item{inp}{- filename of par file}
}
\value{
dataframe corresponding to the par file (or NULL if file does not exist)
}
\description{
Function to create a dataframe from an ADMB par file.
}
\details{
Uses \code{wtsUtilities::selectFile} to open a file dialog if inp is NULL.
}

