% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPar.R
\name{getPar}
\alias{getPar}
\title{Create a TCSAM2013 par object from a par file.}
\usage{
getPar(in.par = NULL)
}
\arguments{
\item{in.par}{= filename of par file}
}
\value{
list object corresponding to the par file
}
\description{
Function to create a TCSAM2013 par object from a par file.
}
\details{

}

