% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRep.R
\name{getRep}
\alias{getRep}
\title{Create a TCSAM2013 rep object from a rep file.}
\usage{
getRep(in.rep = NULL)
}
\arguments{
\item{in.rep}{= filename of rep file}
}
\value{
list object corresponding to the rep file
}
\description{
Function to create a TCSAM2013 rep object from a rep file.
}
\details{

}

