% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRunCommands.R
\name{getRunCommands}
\alias{getRunCommands}
\title{Generate run commands for a tcsam2013 model run}
\usage{
getRunCommands(os = "osx", model = "tcsam2013alta", path2model = "",
  configFile = "", pin = FALSE, hess = FALSE, mcmc = FALSE,
  jitter = FALSE, seed = NULL)
}
\arguments{
\item{os}{- 'win', 'mac' or 'osx'}

\item{model}{- admb model name}

\item{path2model}{- path to model}

\item{configFile}{- filename (including path) to model configuration file}

\item{pin}{- flag (T/F) to use a pin file}

\item{hess}{- flag (T/F) to calculate the hessian}

\item{mcmc}{- flag (T/F) to do mcmc calculations}

\item{jitter}{- flag (T/F) to use jitter initial values}

\item{seed}{- value for random number seed to generate jitter}
}
\description{
Function to generate a script to make a tcsam2013 model run
}
\details{

}

