% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStd.R
\name{getStd}
\alias{getStd}
\title{Create a TCSAM2013 std object from a std file.}
\usage{
getStd(in.std = NULL)
}
\arguments{
\item{in.std}{= filename of std file}
}
\value{
list object corresponding to the std file, or NULL if file does not exist
}
\description{
Function to create a TCSAM2013 std object from a std file.
}
\details{
Uses \code{wtsUtilities::selectFile} to open a file dialog if in.par is NULL.
}

