% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot4.ModelComparisonsGG.TimeSeries.R
\name{plot4.ModelComparisonsGG.TimeSeries}
\alias{plot4.ModelComparisonsGG.TimeSeries}
\title{Plot 4 comparisons of time series from a set of model runs}
\usage{
plot4.ModelComparisonsGG.TimeSeries(dfr, numRecent = 15, facets = NULL,
  scales = "fixed", plotObs = TRUE, plotMod = TRUE, ci = 0.95,
  pdfType = "lognormal", xlab = "year", ylab = NULL, title = NULL,
  xlims = NULL, ylims = NULL, showPlot = TRUE)
}
\arguments{
\item{dfr}{- dataframe}

\item{numRecent}{- number of "recent" years to plot}

\item{facets}{- string giving faceting formula}

\item{scales}{- ggplot2 scales option for facet_grid}

\item{plotObs}{- plot observations}

\item{plotMod}{- plot model fits/predictions}

\item{ci}{- confidence interval for error bars}

\item{pdfType}{- assumed error distribution for confidence intervals}

\item{xlab}{-}

\item{ylab}{-}

\item{title}{-}

\item{xlims}{-}

\item{ylims}{-}

\item{showPlot}{-}
}
\value{
list with requested ggplot objects
}
\description{
Function to plot a comparison of time series from a set of model runs.
}

