% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTCSAM2013.R
\name{plotTCSAM2013}
\alias{plotTCSAM2013}
\title{Plot TCSAM2013 model results.}
\usage{
plotTCSAM2013(obj.rep = NULL, obj.std = NULL, obj.prs = NULL,
  obj.wts = NULL, base.dir = "./", mdl = "tcsam2013alta", endyr = NULL,
  styr = NULL, obsyr = NULL, pltyr = 1965, F35 = 0.73, B35 = 33.54)
}
\arguments{
\item{obj.rep}{- report file list object or filename for R-style report}

\item{obj.std}{- dataframe object with parameter std info or filename for std file}

\item{obj.prs}{- dataframe object w/ parameters info or csv file with parameters info}

\item{obj.wts}{- TCSAM_WTS-type list object}

\item{base.dir}{- path to base folder}

\item{mdl}{- model name (optional if R report file is read)}

\item{endyr}{- assessment year}

\item{styr}{- model start year}

\item{obsyr}{- first year for survey observations}

\item{pltyr}{- first year for some plots}

\item{F35}{- F35 value for control rule plot}

\item{B35}{- B35 value for control rule plot}
}
\value{
list with elements: \cr
\itemize{
 \item rep - obj.rep
 \item std - obj.std
 \item prs - obj.prs
 \item wts - obj.wts
}
}
\description{
Function to plot model results from TCSAM2013.
}
\details{
Convenience wrapper for \code{plotTCSAM2013I()}.
}

