% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTCSAM2013I.GG.R
\name{plotTCSAM2013I.GG}
\alias{plotTCSAM2013I.GG}
\title{Plot TCSAM2013 model results.}
\usage{
plotTCSAM2013I.GG(obj, numRecent = 15, plot1stObs = TRUE, showPlot = TRUE,
  pdf = FALSE)
}
\arguments{
\item{obj}{- object that can be converted into a list of tcsam2013.resLst objects}

\item{numRecent}{- number of "recent" years to include in "zoom" plots}

\item{plot1stObs}{- flag to plot observations from the first case, only}

\item{showPlot}{- flag (T/F) to print plots to screen (if pdf==FALSE)}

\item{pdf}{- flag (T/F) to print plots to a pdf}
}
\value{
list of ggplot2 objects.
}
\description{
Function to plot model results from TCSAM2013.
}
\details{
If pdf is TRUE, the file "TCSAM2013.ModelResults.pdf" is created in the 
working directory.
}
