% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTCSAM2013.R
\name{runTCSAM2013}
\alias{runTCSAM2013}
\title{Function to run TCSAM2013.}
\usage{
runTCSAM2013(os = "osx", path = ".", model = "tcsam2013alta",
  path2model = "", configFile = "", pin = FALSE, hess = FALSE,
  mcmc = FALSE, minPhase = NULL, maxPhase = NULL, jitter = FALSE,
  seed = NULL, plotResults = hess, rmEXE = FALSE, cleanup = TRUE)
}
\arguments{
\item{os}{- 'win' or 'mac' or 'osx'}

\item{path}{- path for model output}

\item{model}{- TCSAM2013 model executable name}

\item{path2model}{- path to model executable}

\item{configFile}{- filename (including path) to model configuration file}

\item{pin}{- flag (T/F) to use/not use a default pin file, or an ascii filename to use a non-default one}

\item{hess}{- T/F to compute hessian (and .std file)}

\item{mcmc}{- T/F to run mcmc}

\item{minPhase}{- min phase to start estimation}

\item{maxPhase}{- max phase for estimation}

\item{jitter}{- T/F to jitter parameters}

\item{seed}{- seed for random number generator (or NULL)}

\item{plotResults}{- T/F to plot results using \code{plotTCSAM2013I.GG}}

\item{rmEXE}{- flag to remove model executable from run folder}

\item{cleanup}{- flag (T/F) to clean up unnecessary files}
}
\value{
- dataframe with 2 columns (name, value) with jitter seed (if jittered) and par file info, or NULL if par file does not exist
}
\description{
This functions runs a TCSAM2013 model once.
}
\details{
This function creates a shell script ('./tmp.sh') in the
working directory and uses it to run a version of the TCSAM2013 model.\cr
Initial model parameters can be jittered based on the system clock time as a seed
to the random number generator. The seed and final objective function value are
saved for each model run in a csv file (the value of out.csv).


If the path associated with \code{configFile} is a relative one, it should
be relative to the \code{path} variable.
}
